move fs C stub to separated file
