+++ src/internal/event_loop/thread_pool.c
@@ -35,25 +35,20 @@
 #include <unistd.h>
 #include <stdlib.h>
 #include <fcntl.h>
-#include <stdio.h>
 #include <string.h>
 #include <errno.h>
 #include <time.h>
-#include <dirent.h>
 #if !defined(__ANDROID__) || __ANDROID_API__ >= 28
 #include <spawn.h>
 #endif
 #include <sys/socket.h>
 #include <arpa/inet.h>
 #include <netdb.h>
-#include <sys/stat.h>
-#include <sys/file.h>
 #include <sys/wait.h>
 #include <poll.h>
 
 #ifdef __linux__
 #include <sys/syscall.h>
-#include <linux/fs.h>
 #include <sys/inotify.h>
 #include <linux/version.h>
 #endif
@@ -533,14 +528,14 @@ int32_t moonbitlang_async_errno_is_cancelled(int32_t err) {
 // ===================== concrete jobs =====================
 // =========================================================
 
-static
-void *make_job(
+MOONBIT_FFI_EXPORT
+void *moonbitlang_async_make_job(
   int32_t size,
   void (*free)(void*),
   int32_t (*worker)(void*, int32_t*),
   int32_t (*cancel_handler)(void*)
 ) {
-  struct job *job = (struct job*)malloc(size);
+  struct job *job = (struct job*)malloc(size + sizeof(struct job));
   job->ret = 0;
   job->err = 0;
   job->worker = worker;
@@ -549,8 +544,8 @@ void *make_job(
   return job->payload;
 }
 
-#define MAKE_JOB(name, cancel_handler) (struct name##_job*)make_job(\
-  sizeof(struct job) + sizeof(struct name##_job),\
+#define MAKE_JOB(name, cancel_handler) (struct name##_job*)moonbitlang_async_make_job(\
+  sizeof(struct name##_job),\
   (void (*)(void*))free_##name##_job,\
   (int32_t (*)(void*, int32_t*)) name##_job_worker,\
   (int32_t (*)(void*))cancel_handler\
@@ -780,1274 +775,6 @@ struct write_job *moonbitlang_async_make_write_job(
   job->len = len;
   job->position = position;
   return job;
-}
-
-// ===== open job =====
-
-struct open_job {
-  char *filename;
-  int access;
-  int create_mode;
-  int append;
-  int sync;
-  int mode;
-  HANDLE result;
-#ifdef _WIN32
-  BY_HANDLE_FILE_INFORMATION stat;
-#else
-  struct stat stat;
-#endif
-};
-
-static
-void free_open_job(struct open_job *job) {
-  moonbit_decref(job->filename);
-}
-
-static
-int32_t open_job_worker(struct open_job *job, int32_t *err_out) {
-#ifdef _WIN32
-  static int access_flags[] = {
-    GENERIC_READ,
-    GENERIC_WRITE,
-    GENERIC_READ | GENERIC_WRITE,
-    FILE_LIST_DIRECTORY
-  };
-  static int create_modes[] = { OPEN_EXISTING, TRUNCATE_EXISTING, OPEN_ALWAYS, CREATE_ALWAYS, CREATE_NEW };
-  static int sync_flags[] = { 0, FILE_FLAG_WRITE_THROUGH, FILE_FLAG_WRITE_THROUGH };
-#else
-  static int access_flags[] = { O_RDONLY, O_WRONLY, O_RDWR, O_RDONLY };
-  static int create_modes[] = {
-    0,
-    O_TRUNC,
-    O_CREAT,
-    O_CREAT | O_TRUNC,
-    O_CREAT | O_EXCL
-  };
-  static int sync_flags[] = { 0, O_DSYNC, O_SYNC };
-#endif
-
-#ifdef _WIN32
-
-  DWORD flags =
-    FILE_ATTRIBUTE_NORMAL
-    | FILE_FLAG_BACKUP_SEMANTICS
-    | sync_flags[job->sync];
-
-  if (job->access == 3)
-    flags |= FILE_FLAG_OVERLAPPED;
-
-  DWORD access_flag = access_flags[job->access];
-  if (job->append)
-    access_flag = (access_flag ^ GENERIC_WRITE) | FILE_APPEND_DATA;
-
-  while (1) {
-    job->result = CreateFileW(
-      (LPCWSTR)job->filename,
-      access_flag, // desired access
-      FILE_SHARE_READ | FILE_SHARE_WRITE | FILE_SHARE_DELETE, // shared mode
-      NULL, // security attributes
-      create_modes[job->create_mode], // creation
-      flags, // flags and attributes. Note that we open files in synchronous mode
-      NULL // template file
-    );
-
-    if (job->result != INVALID_HANDLE_VALUE)
-      break;
-
-    // handle error
-    int err = GetLastError();
-    if (err != ERROR_PIPE_BUSY) {
-      *err_out = err;
-      return -1;
-    }
-
-    // We are trying to open a named pipe, but no pipe instance is available,
-    // so wait until any instance is available.
-    // This wait is cancellable via `CancelSynchronousIo`.
-    if (!WaitNamedPipeW((LPCWSTR)job->filename, NMPWAIT_WAIT_FOREVER)) {
-      *err_out = GetLastError();
-      return -1;
-    }
-  }
-
-  // get the kind of the file
-  if (!GetFileInformationByHandle(job->result, &job->stat)) {
-    *err_out = GetLastError();
-    CloseHandle(job->result);
-    return -1;
-  }
-
-#else
-
-  int flags =
-    access_flags[job->access]
-    | sync_flags[job->sync]
-    | create_modes[job->create_mode];
-  if (job->append) flags |= O_APPEND;
-
-  job->result = open(
-    job->filename,
-    flags | O_CLOEXEC,
-    job->mode
-  );
-  if (job->result < 0) {
-    *err_out = errno;
-    return -1;
-  }
-
-  if (fstat(job->result, &job->stat) < 0) {
-    *err_out = errno;
-    close(job->result);
-    return -1;
-  }
-
-#endif
-  return 0;
-}
-
-MOONBIT_FFI_EXPORT
-struct open_job *moonbitlang_async_make_open_job(
-  char *filename,
-  int access,
-  int create_mode,
-  int append,
-  int sync,
-  int mode
-) {
-
-  struct open_job *job = MAKE_JOB(open, 0);
-  job->filename = filename;
-  job->access = access;
-  job->create_mode = create_mode;
-  job->append = append;
-  job->sync = sync;
-  job->mode = mode;
-  return job;
-}
-
-MOONBIT_FFI_EXPORT
-HANDLE moonbitlang_async_open_job_get_fd(struct open_job *job) {
-  return job->result;
-}
-
-MOONBIT_FFI_EXPORT
-int32_t moonbitlang_async_open_job_get_kind(struct open_job *job) {
-#ifdef _WIN32
-  return moonbitlang_async_kind_from_attr(job->stat.dwFileAttributes);
-#else
-  return moonbitlang_async_file_kind_from_stat(&job->stat);
-#endif
-}
-
-MOONBIT_FFI_EXPORT
-uint64_t moonbitlang_async_open_job_get_dev_id(struct open_job *job) {
-#ifdef _WIN32
-  return job->stat.dwVolumeSerialNumber;
-#else
-  return job->stat.st_dev;
-#endif
-}
-
-MOONBIT_FFI_EXPORT
-uint64_t moonbitlang_async_open_job_get_file_id(struct open_job *job) {
-#ifdef _WIN32
-  return ((uint64_t)(job->stat.nFileIndexHigh) << 32) | (uint64_t)(job->stat.nFileIndexLow);
-#else
-  return job->stat.st_ino;
-#endif
-}
-
-// ===== file kind of fd job, get kind of an existing fd =====
-struct kind_of_fd_job {
-  HANDLE fd;
-};
-
-static
-void free_kind_of_fd_job(struct kind_of_fd_job *job) {}
-
-static
-int32_t kind_of_fd_job_worker(struct kind_of_fd_job *job, int32_t *err_out) {
-  int32_t ret = moonbitlang_async_kind_of_fd(job->fd);
-  if (ret < 0)
-    *err_out = GetLastError();
-  return ret;
-}
-
-struct kind_of_fd_job *moonbitlang_async_make_kind_of_fd_job(HANDLE fd) {
-  struct kind_of_fd_job *job = MAKE_JOB(kind_of_fd, 0);
-  job->fd = fd;
-  return job;
-}
-
-// ===== file kind by path job, get kind of path on file system =====
-
-struct file_kind_by_path_job {
-  HANDLE parent;
-  char *path;
-  int follow_symlink;
-};
-
-static
-void free_file_kind_by_path_job(struct file_kind_by_path_job *job) {
-  moonbit_decref(job->path);
-}
-
-static
-int32_t file_kind_by_path_job_worker(struct file_kind_by_path_job *job, int32_t *err_out) {
-#ifdef _WIN32
-  DWORD flags = FILE_ATTRIBUTE_NORMAL | FILE_FLAG_BACKUP_SEMANTICS;
-  if (!job->follow_symlink)
-    flags |= FILE_FLAG_OPEN_REPARSE_POINT;
-
-  HANDLE handle = CreateFileW(
-    (LPCWSTR)job->path,
-    FILE_READ_ATTRIBUTES, // desired access
-    FILE_SHARE_READ | FILE_SHARE_WRITE | FILE_SHARE_DELETE, // shared mode
-    NULL, // security attributes
-    OPEN_EXISTING, // creation mode
-    flags, // flags and attributes
-    NULL // template file
-  );
-
-  if (handle == INVALID_HANDLE_VALUE) {
-    *err_out = GetLastError();
-    return -1;
-  }
-
-  int32_t ret = moonbitlang_async_kind_of_fd(handle);
-  if (ret < 0)
-    *err_out = GetLastError();
-
-  CloseHandle(handle);
-  return ret;
-
-#else
-
-  struct stat stat_obj;
-  int ret = fstatat(
-    job->parent < 0 ? AT_FDCWD : job->parent,
-    job->path,
-    &stat_obj,
-    job->follow_symlink ? 0 : AT_SYMLINK_NOFOLLOW
-  );
-  if (ret < 0) {
-    *err_out = errno;
-    return -1;
-  } else {
-    return moonbitlang_async_file_kind_from_stat(&stat_obj);
-  }
-
-#endif
-}
-
-struct file_kind_by_path_job *moonbitlang_async_make_file_kind_by_path_job(
-  HANDLE parent,
-  char *path,
-  int follow_symlink
-) {
-  struct file_kind_by_path_job *job = MAKE_JOB(file_kind_by_path, 0);
-  job->parent = parent;
-  job->path = path;
-  job->follow_symlink = follow_symlink;
-  return job;
-}
-
-// ===== file size job, get size of opened file =====
-
-struct file_size_job {
-  HANDLE fd;
-  int64_t result;
-};
-
-static
-void free_file_size_job(struct file_size_job *job) {}
-
-static
-int32_t file_size_job_worker(struct file_size_job *job, int32_t *err_out) {
-#ifdef _WIN32
-  LARGE_INTEGER size;
-  if (!GetFileSizeEx(job->fd, &size)) {
-    *err_out = GetLastError();
-    return -1;
-  }
-  job->result = size.QuadPart;
-#else
-  struct stat stat;
-  int32_t ret = fstat(job->fd, &stat);
-  if (ret < 0) {
-    *err_out = errno;
-    return ret;
-  } else {
-    job->result = stat.st_size;
-  }
-#endif
-  return 0;
-}
-
-struct file_size_job *moonbitlang_async_make_file_size_job(HANDLE fd) {
-  struct file_size_job *job = MAKE_JOB(file_size, 0);
-  job->fd = fd;
-  return job;
-}
-
-int64_t moonbitlang_async_get_file_size_result(struct file_size_job *job) {
-  return job->result;
-}
-
-// ===== file time job, get timestamp of opened file =====
-
-struct file_time_job {
-  HANDLE fd;
-  void *out;
-};
-
-static
-void free_file_time_job(struct file_time_job *job) {
-  moonbit_decref(job->out);
-}
-
-static
-int32_t file_time_job_worker(struct file_time_job *job, int32_t *err_out) {
-#ifdef _WIN32
-  if (
-    !GetFileInformationByHandleEx(
-      job->fd,
-      FileBasicInfo,
-      job->out,
-      sizeof(FILE_BASIC_INFO)
-    )
-  ) {
-    *err_out = GetLastError();
-    return -1;
-  }
-#else
-  int32_t ret = fstat(job->fd, job->out);
-  if (ret < 0)
-    *err_out = errno;
-  return ret;
-#endif
-  return 0;
-}
-
-struct file_time_job *moonbitlang_async_make_file_time_job(HANDLE fd, void *out) {
-  struct file_time_job *job = MAKE_JOB(file_time, 0);
-  job->fd = fd;
-  job->out = out;
-  return job;
-}
-
-// ===== file time by path job, get timestamp of path on file system =====
-
-struct file_time_by_path_job {
-  char *path;
-  void *out;
-  int follow_symlink;
-};
-
-static
-void free_file_time_by_path_job(struct file_time_by_path_job *job) {
-  moonbit_decref(job->path);
-  moonbit_decref(job->out);
-}
-
-static
-int32_t file_time_by_path_job_worker(struct file_time_by_path_job *job, int32_t *err_out) {
-#ifdef _WIN32
-  DWORD flags = FILE_ATTRIBUTE_NORMAL | FILE_FLAG_BACKUP_SEMANTICS;
-  if (!job->follow_symlink)
-    flags |= FILE_FLAG_OPEN_REPARSE_POINT;
-
-  HANDLE handle = CreateFileW(
-    (LPCWSTR)job->path,
-    FILE_READ_ATTRIBUTES, // desired access
-    FILE_SHARE_READ | FILE_SHARE_WRITE | FILE_SHARE_DELETE, // shared mode
-    NULL, // security attributes
-    OPEN_EXISTING, // creation mode
-    flags, // flags and attributes
-    NULL // template file
-  );
-  if (handle == INVALID_HANDLE_VALUE) {
-    *err_out = GetLastError();
-    return -1;
-  }
-
-  if (
-    !GetFileInformationByHandleEx(
-      handle,
-      FileBasicInfo,
-      job->out,
-      sizeof(FILE_BASIC_INFO)
-    )
-  ) {
-    *err_out = GetLastError();
-    CloseHandle(handle);
-    return -1;
-  }
-
-  CloseHandle(handle);
-  return 0;
-
-#else
-
-  int32_t ret;
-  if (job->follow_symlink) {
-    ret = stat(job->path, job->out);
-  } else {
-    ret = lstat(job->path, job->out);
-  }
-  if (ret < 0)
-    *err_out = errno;
-  return ret;
-
-#endif
-}
-
-struct file_time_by_path_job *moonbitlang_async_make_file_time_by_path_job(
-  char *path,
-  void *out,
-  int follow_symlink
-) {
-  struct file_time_by_path_job *job = MAKE_JOB(file_time_by_path, 0);
-  job->path = path;
-  job->out = out;
-  job->follow_symlink = follow_symlink;
-  return job;
-}
-
-// ===== chmod job, change permission of file =====
-
-struct chmod_job {
-  char *path;
-  int mode;
-};
-
-static
-void free_chmod_job(struct chmod_job *job) {
-  moonbit_decref(job->path);
-}
-
-static
-int32_t chmod_job_worker(struct chmod_job *job, int32_t *err_out) {
-#ifdef _WIN32
-
-  *err_out = ERROR_NOT_SUPPORTED;
-  return -1;
-
-#else
-
-  int32_t ret = chmod(job->path, job->mode);
-  if (ret < 0)
-    *err_out = errno;
-  return ret;
-
-#endif
-}
-
-struct chmod_job *moonbitlang_async_make_chmod_job(char *path, int mode) {
-  struct chmod_job *job = MAKE_JOB(chmod, 0);
-  job->path = path;
-  job->mode = mode;
-  return job;
-}
-
-// ===== fsync job, synchronize file modification to disk =====
-
-struct fsync_job {
-  HANDLE fd;
-  int only_data;
-};
-
-static
-void free_fsync_job(struct fsync_job *job) {}
-
-static
-int32_t fsync_job_worker(struct fsync_job *job, int32_t *err_out) {
-#ifdef _WIN32
-
-  if (!FlushFileBuffers(job->fd)) {
-    *err_out = GetLastError();
-    return -1;
-  }
-  return 0;
-
-#else
-
-#if defined(__MACH__)
-  // it seems that `fdatasync` is not available on some MacOS versions
-  int32_t ret = fsync(job->fd);
-#else
-  int32_t ret;
-  if (job->only_data) {
-    ret = fdatasync(job->fd);
-  } else {
-    ret = fsync(job->fd);
-  }
-#endif
-
-  if (ret < 0)
-    *err_out = errno;
-  return ret;
-
-#endif
-}
-
-struct fsync_job *moonbitlang_async_make_fsync_job(HANDLE fd, int only_data) {
-  struct fsync_job *job = MAKE_JOB(fsync, 0);
-  job->fd = fd;
-  job->only_data = only_data;
-  return job;
-}
-
-// ===== flock job, place advisory lock on a file =====
-struct flock_job {
-  HANDLE fd;
-  int exclusive;
-};
-
-static
-void free_flock_job(struct flock_job *job) {}
-
-static
-int32_t flock_job_worker(struct flock_job *job, int32_t *err_out) {
-#ifdef _WIN32
-
-  OVERLAPPED overlapped;
-  memset(&overlapped, 0, sizeof(OVERLAPPED));
-  // We want to provide advisory lock here
-  // (i.e. only lock operations conflict with each other, raw IO are not affected),
-  // because mandatory file lock is not available on Linux/MacOS.
-  // However, Windows only provides mandatory file lock.
-  // Fortunately, https://learn.microsoft.com/en-us/windows/win32/api/fileapi/nf-fileapi-lockfileex
-  // explicitly state that locking a region beyond end of file is *not* an error.
-  // So, here we lock the last byte in the whole address space to simulate advisory locking,
-  // as this region can almost never get touched by normal IO operations.
-  overlapped.Offset = 0xfffffffe;
-  overlapped.OffsetHigh = 0xffffffff;
-  BOOL ret = LockFileEx(
-    job->fd,
-    job->exclusive ? LOCKFILE_EXCLUSIVE_LOCK : 0,
-    0, // reserved
-    1,
-    0,
-    &overlapped
-  );
-
-  if (!ret) {
-    *err_out = GetLastError();
-    return -1;
-  }
-  return 0;
-
-#else
-
-  int32_t ret = flock(job->fd, job->exclusive ? LOCK_EX : LOCK_SH);
-  if (ret < 0)
-    *err_out = errno;
-  return ret;
-
-#endif
-}
-
-struct flock_job *moonbitlang_async_make_flock_job(HANDLE fd, int exclusive) {
-  struct flock_job *job = MAKE_JOB(flock, 0);
-  job->fd = fd;
-  job->exclusive = exclusive;
-  return job;
-}
-
-// ===== remove job, remove file from file system =====
-
-struct remove_job {
-  char *path;
-};
-
-static
-void free_remove_job(struct remove_job *job) {
-  moonbit_decref(job->path);
-}
-
-static
-int32_t remove_job_worker(struct remove_job *job, int32_t *err_out) {
-#ifdef _WIN32
-  DWORD attrs = GetFileAttributesW((LPCWSTR)job->path);
-  if (attrs == INVALID_FILE_ATTRIBUTES) {
-    *err_out = GetLastError();
-    return -1;
-  }
-
-  BOOL ret;
-  // Simulate POSIX behavior on Windows.
-  // Maybe we should just merge `@fs.remove` and `@fs.rmdir`?
-  if ((attrs & FILE_ATTRIBUTE_DIRECTORY) && (attrs & FILE_ATTRIBUTE_REPARSE_POINT)) {
-    ret = RemoveDirectoryW((LPCWSTR)job->path);
-  } else {
-    ret = DeleteFileW((LPCWSTR)job->path);
-  }
-
-  if (!ret) {
-    *err_out = GetLastError();
-    return -1;
-  }
-  return 0;
-#else
-  int32_t ret = remove(job->path);
-  if (ret < 0)
-    *err_out = errno;
-  return ret;
-#endif
-}
-
-struct remove_job *moonbitlang_async_make_remove_job(char *path) {
-  struct remove_job *job = MAKE_JOB(remove, 0);
-  job->path = path;
-  return job;
-}
-
-// ===== access job, test permission of file path =====
-
-struct access_job {
-  char *path;
-  int amode;
-};
-
-static
-void free_access_job(struct access_job *job) {
-  moonbit_decref(job->path);
-}
-
-static
-int32_t access_job_worker(struct access_job *job, int32_t *err_out) {
-#ifdef _WIN32
-
-  static int access_modes[] = { 0, GENERIC_READ, GENERIC_WRITE, FILE_EXECUTE };
-
-  HANDLE handle = CreateFileW(
-    (LPCWSTR)job->path,
-    access_modes[job->amode],
-    FILE_SHARE_DELETE | FILE_SHARE_READ | FILE_SHARE_WRITE,
-    NULL,
-    OPEN_EXISTING,
-    FILE_ATTRIBUTE_NORMAL | FILE_FLAG_BACKUP_SEMANTICS,
-    NULL
-  );
-  if (handle == INVALID_HANDLE_VALUE) {
-    *err_out = GetLastError();
-    return -1;
-  } else {
-    CloseHandle(handle);
-    return 0;
-  }
-
-#else
-
-  static int access_modes[] = { F_OK, R_OK, W_OK, X_OK };
-  int32_t ret = access(job->path, access_modes[job->amode]);
-  if (ret < 0)
-    *err_out = errno;
-  return ret;
-
-#endif
-}
-
-struct access_job *moonbitlang_async_make_access_job(char *path, int amode) {
-  struct access_job *job = MAKE_JOB(access, 0);
-  job->path = path;
-  job->amode = amode;
-  return job;
-}
-
-// ===== rename job, rename file =====
-struct rename_job {
-  char *old_path;
-  char *new_path;
-  int32_t replace;
-};
-
-static
-void free_rename_job(struct rename_job *job) {
-  moonbit_decref(job->old_path);
-  moonbit_decref(job->new_path);
-}
-
-static
-int32_t rename_job_worker(struct rename_job *job, int32_t *err_out) {
-#ifdef _WIN32
-
-  HANDLE handle = CreateFileW(
-    (LPCWSTR)job->old_path,
-    DELETE,
-    FILE_SHARE_DELETE | FILE_SHARE_READ | FILE_SHARE_WRITE,
-    NULL,
-    OPEN_EXISTING,
-    FILE_ATTRIBUTE_NORMAL | FILE_FLAG_BACKUP_SEMANTICS,
-    NULL
-  );
-
-  if (handle == INVALID_HANDLE_VALUE) {
-    *err_out = GetLastError();
-    return -1;
-  }
-
-  int new_path_len = Moonbit_array_length(job->new_path);
-  int buffer_size = sizeof(FILE_RENAME_INFO) + new_path_len * 2 + 2;
-  FILE_RENAME_INFO *info = (FILE_RENAME_INFO*)malloc(buffer_size);
-
-  // 3 = FILE_RENAME_REPLACE_IF_EXISTS | FILE_RENAME_POSIX_SEMANTICS 
-  info->Flags = job->replace ? 3 : 0;
-  info->RootDirectory = NULL;
-  info->FileNameLength = new_path_len * 2;
-  memcpy(info->FileName, job->new_path, new_path_len * 2);
-  info->FileName[new_path_len] = 0;
-
-  BOOL ret = SetFileInformationByHandle(handle, FileRenameInfoEx, info, buffer_size);
-
-  CloseHandle(handle);
-  free(info);
-
-  if (ret)
-    return 0;
-
-  if (GetLastError() != ERROR_INVALID_PARAMETER) {
-    *err_out = GetLastError();
-    return -1;
-  }
-
-  // fallback on older systems
-
-  ret = MoveFileExW(
-    (LPCWSTR)job->old_path,
-    (LPCWSTR)job->new_path,
-    MOVEFILE_COPY_ALLOWED | (job->replace ? MOVEFILE_REPLACE_EXISTING : 0)
-  );
-  if (!ret) {
-    *err_out = GetLastError();
-    return -1;
-  }
-  return 0;
-
-#elif defined(__MACH__)
-
-  int32_t ret = renameatx_np(
-    AT_FDCWD, job->old_path,
-    AT_FDCWD, job->new_path,
-    job->replace ? 0 : RENAME_EXCL
-  );
-  if (ret < 0)
-    *err_out = errno;
-  return ret;
-
-#elif defined(__linux__)
-
-  int32_t ret = syscall(
-    SYS_renameat2,
-    AT_FDCWD, job->old_path,
-    AT_FDCWD, job->new_path,
-    job->replace ? 0 : RENAME_NOREPLACE
-  );
-  if (ret < 0)
-    *err_out = errno;
-  return ret;
-
-#else
-
-  *err_out = ENOSYS;
-  return -1;
-
-#endif
-}
-
-struct rename_job *moonbitlang_async_make_rename_job(
-  char *old_path,
-  char *new_path,
-  int32_t replace
-) {
-  struct rename_job *job = MAKE_JOB(rename, 0);
-  job->old_path = old_path;
-  job->new_path = new_path;
-  job->replace = replace;
-  return job;
-}
-
-// ===== symlink job, create symbolic link =====
-
-struct symlink_job {
-  char *target;
-  char *path;
-  int32_t force_symlink;
-};
-
-static
-void free_symlink_job(struct symlink_job *job) {
-  moonbit_decref(job->target);
-  moonbit_decref(job->path);
-}
-
-#ifdef _WIN32
-typedef struct {
-    USHORT SubstituteNameOffset;
-    USHORT SubstituteNameLength;
-    USHORT PrintNameOffset;
-    USHORT PrintNameLength;
-    WCHAR  PathBuffer[1];
-} MOUNT_POINT_REPARSE_BUFFER;
-
-typedef struct {
-  ULONG ReparseTag;
-  USHORT ReparseDataLength;
-  USHORT Reserved;
-  MOUNT_POINT_REPARSE_BUFFER MountPointReparseBuffer;
-} REPARSE_DATA_BUFFER;
-#endif
-
-static
-int32_t symlink_job_worker(struct symlink_job *job, int32_t *err_out) {
-#ifdef _WIN32
-
-  int target_len = Moonbit_array_length(job->target);
-  LPCWSTR target = (LPCWSTR)job->target;
-  LPCWSTR path = (LPCWSTR)job->path;
-
-  DWORD attrs = GetFileAttributesW(target);
-  BOOL is_dir = attrs != INVALID_FILE_ATTRIBUTES && (attrs & FILE_ATTRIBUTE_DIRECTORY);
-
-  // create NTFS junction if possible
-  if (job->force_symlink)
-    goto symlink_fallback;
-
-  if (!is_dir)
-    goto symlink_fallback;
-
-  if (wcsncmp(target, L"\\??\\", 4) == 0 || wcsncmp(target, L"\\\\?\\", 4) == 0) {
-    target += 4;
-    target_len -= 4;
-  }
-
-  if (
-    target_len >= 3
-    && ('a' <= target[0] && target[0] <= 'z' || 'A' <= target[0] && target[0] <= 'Z')
-    && target[1] == ':'
-    && (target[2] == '\\' || target[2] == '/')
-  ) {
-    // normal absolute path
-  } else if (wcsncmp(target, L"\\", 2) == 0) {
-    // UNC path for network resource, does not support junction
-    goto symlink_fallback;
-  } else {
-    // relaive path, does not support junction
-    goto symlink_fallback;
-  }
-
-  if (!CreateDirectoryW(path, NULL)) {
-    *err_out = GetLastError();
-    return -1;
-  }
-
-  HANDLE link = CreateFileW(
-    path,
-    GENERIC_WRITE,
-    0,
-    NULL,
-    OPEN_EXISTING,
-    FILE_FLAG_OPEN_REPARSE_POINT | FILE_FLAG_BACKUP_SEMANTICS,
-    NULL
-  );
-  if (link == INVALID_HANDLE_VALUE) {
-    *err_out = GetLastError();
-    RemoveDirectoryW(path);
-    return -1;
-  }
-
-  DWORD substitute_name_len =
-    2 * target_len
-    + 8 // NT path "\??\" prefix for substitute path
-  ;
-  DWORD print_name_len = 2 * target_len;
-
-  DWORD path_buffer_length =
-    substitute_name_len
-    + print_name_len
-    + 4 // NUL terminator for substitute path and print path
-  ;
-
-  DWORD reparse_buffer_length =
-    offsetof(MOUNT_POINT_REPARSE_BUFFER, PathBuffer)
-    + path_buffer_length;
-
-  DWORD buffer_size =
-    offsetof(REPARSE_DATA_BUFFER, MountPointReparseBuffer)
-    + reparse_buffer_length;
-
-  REPARSE_DATA_BUFFER *buf = (REPARSE_DATA_BUFFER*)malloc(buffer_size);
-  buf->ReparseTag = IO_REPARSE_TAG_MOUNT_POINT;
-  buf->ReparseDataLength = reparse_buffer_length;
-  buf->Reserved = 0;
-  buf->MountPointReparseBuffer.SubstituteNameOffset = 0;
-  buf->MountPointReparseBuffer.SubstituteNameLength = substitute_name_len;
-  memcpy(
-    buf->MountPointReparseBuffer.PathBuffer,
-    L"\\??\\",
-    8
-  );
-  memcpy(
-    buf->MountPointReparseBuffer.PathBuffer + 4,
-    target,
-    target_len * 2 + 2
-  );
-  for (WCHAR *ptr = buf->MountPointReparseBuffer.PathBuffer + 4; *ptr; ++ptr) {
-    // substitute path does not support forward slash
-    if (*ptr == L'/')
-      *ptr = L'\\';
-  }
-  buf->MountPointReparseBuffer.PrintNameOffset = substitute_name_len + 2;
-  buf->MountPointReparseBuffer.PrintNameLength = print_name_len;
-  memcpy(
-    buf->MountPointReparseBuffer.PathBuffer + 5 + target_len,
-    // avoid substituting forward slash twice, reuse the substituted result
-    buf->MountPointReparseBuffer.PathBuffer + 4,
-    target_len * 2 + 2
-  );
-
-  DWORD bytes_returned = 0;
-  BOOL ok = DeviceIoControl(
-    link,
-    FSCTL_SET_REPARSE_POINT,
-    buf,
-    buffer_size,
-    NULL,
-    0,
-    &bytes_returned,
-    NULL
-  );
-  int err = GetLastError();
-  free(buf);
-  CloseHandle(link);
-
-  if (!ok) {
-    RemoveDirectoryW(path);
-    if (err == ERROR_INVALID_PARAMETER) {
-      // this is mainly for handling non-NTFS volume.
-      // There are other cases that will also generate `ERROR_INVALID_PARAMETER`,
-      // such as invalid character in path.
-      // But for those cases, the symlink fallback path will give a similar error,
-      // so the net result is still the same.
-      goto symlink_fallback;
-    }
-    *err_out = err;
-    return -1;
-  }
-
-  return 0;
-
-symlink_fallback:
-
-  if (
-    !CreateSymbolicLinkW(
-      (LPCWSTR)job->path,
-      (LPCWSTR)job->target,
-      is_dir ? SYMBOLIC_LINK_FLAG_DIRECTORY : 0
-    )
-  ) {
-    *err_out = GetLastError();
-    return -1;
-  }
-  return 0;
-
-#else
-
-  int32_t ret = symlink(job->target, job->path);
-  if (ret < 0)
-    *err_out = errno;
-  return ret;
-
-#endif
-}
-
-struct symlink_job *moonbitlang_async_make_symlink_job(
-  char *target,
-  char *path,
-  int32_t force_symlink
-) {
-  struct symlink_job *job = MAKE_JOB(symlink, 0);
-  job->target = target;
-  job->path = path;
-  job->force_symlink = force_symlink;
-  return job;
-}
-
-// ===== mkdir job, create new directory =====
-
-struct mkdir_job {
-  char *path;
-  int mode;
-};
-
-static
-void free_mkdir_job(struct mkdir_job *job) {
-  moonbit_decref(job->path);
-}
-
-static
-int32_t mkdir_job_worker(struct mkdir_job *job, int32_t *err_out) {
-#ifdef _WIN32
-
-  if (!CreateDirectoryW((LPCWSTR)job->path, NULL)) {
-    *err_out = GetLastError();
-    return -1;
-  }
-  return 0;
-
-#else
-
-  int32_t ret = mkdir(job->path, job->mode);
-  if (ret < 0)
-    *err_out = errno;
-  return ret;
-
-#endif
-}
-
-struct mkdir_job *moonbitlang_async_make_mkdir_job(char *path, int mode) {
-  struct mkdir_job *job = MAKE_JOB(mkdir, 0);
-  job->path = path;
-  job->mode = mode;
-  return job;
-}
-
-// ===== rmdir job, remove directory =====
-
-struct rmdir_job {
-  char *path;
-};
-
-static
-void free_rmdir_job(struct rmdir_job *job) {
-  moonbit_decref(job->path);
-}
-
-static
-int32_t rmdir_job_worker(struct rmdir_job *job, int32_t *err_out) {
-#ifdef _WIN32
-
-  if (!RemoveDirectoryW((LPCWSTR)job->path)) {
-    *err_out = GetLastError();
-    return -1;
-  }
-  return 0;
-
-#else
-
-  int32_t ret = rmdir(job->path);
-  if (ret < 0)
-    *err_out = errno;
-  return ret;
-
-#endif
-}
-
-struct rmdir_job *moonbitlang_async_make_rmdir_job(char *path) {
-  struct rmdir_job *job = MAKE_JOB(rmdir, 0);
-  job->path = path;
-  return job;
-}
-
-// ===== readdir job, read directory entry =====
-
-struct readdir_job {
-  HANDLE dir;
-  void *out;
-  int32_t len;
-  int32_t restart;
-};
-
-static
-void free_readdir_job(struct readdir_job *job) {}
-
-static
-int32_t readdir_job_worker(struct readdir_job *job, int32_t *err_out) {
-#ifdef _WIN32
-  if (
-    !GetFileInformationByHandleEx(
-      job->dir,
-      job->restart ?  FileIdBothDirectoryRestartInfo : FileIdBothDirectoryInfo,
-      job->out,
-      job->len
-    )
-  ) {
-    if (GetLastError() == ERROR_NO_MORE_FILES)
-      return 0;
-    else {
-      *err_out = GetLastError();
-      return -1;
-    }
-  }
-
-  // `GetFileInformationByHandleEx` does not support a total length
-  return job->len;
-
-#elif defined(__linux__)
-
-  if (job->restart && lseek(job->dir, 0, SEEK_SET) < 0) {
-    *err_out = errno;
-    return -1;
-  }
-
-  int32_t ret = syscall(SYS_getdents64, job->dir, job->out, job->len);
-  if (ret < 0)
-    *err_out = errno;
-  return ret;
-
-#elif defined(__MACH__)
-
-  if (job->restart && lseek(job->dir, 0, SEEK_SET) < 0) {
-    *err_out = errno;
-    return -1;
-  }
-
-  struct attrlist attr_spec = {
-    ATTR_BIT_MAP_COUNT,
-    0, // reserved
-    ATTR_CMN_NAME | ATTR_CMN_RETURNED_ATTRS | ATTR_CMN_OBJTYPE | ATTR_CMN_FILEID, // commonattr
-    0, // volattr
-    0, // dirattr
-    0, // fileattr
-    0 // forkattr
-  };
-  int32_t ret = getattrlistbulk(job->dir, &attr_spec, job->out, job->len, 0);
-  if (ret < 0)
-    *err_out = errno;
-  return ret;
-
-#else
-
-  *err_out = ENOSYS;
-  return -1;
-
-#endif
-}
-
-struct readdir_job *moonbitlang_async_make_readdir_job(
-  HANDLE dir,
-  void *out,
-  int32_t len,
-  int32_t restart
-) {
-  struct readdir_job *job = MAKE_JOB(readdir, 0);
-  job->dir = dir;
-  job->out = out;
-  job->len = len;
-  job->restart = restart;
-  return job;
-}
-
-// ===== realpath job, get canonical representation of a path =====
-
-#ifdef _WIN32
-#define REALPATH_JOB_BUFFER_LENGTH 1024
-#endif
-
-struct realpath_job {
-  char *path;
-  char *result;
-#ifdef _WIN32
-  // avoid some allocation in the simple case
-  WCHAR buf[REALPATH_JOB_BUFFER_LENGTH];
-#endif
-};
-
-static
-void free_realpath_job(struct realpath_job *job) {
-  moonbit_decref(job->path);
-#ifdef _WIN32
-  if (job->result && job->result != (char*)job->buf)
-    free(job->result);
-#else
-  if (job->result)
-    free(job->result);
-#endif
-}
-
-static
-int32_t realpath_job_worker(struct realpath_job *job, int32_t *err_out) {
-#ifdef _WIN32
-  HANDLE file = CreateFileW(
-    (LPCWSTR)job->path,
-    0,
-    FILE_SHARE_READ | FILE_SHARE_WRITE | FILE_SHARE_DELETE,
-    NULL,
-    OPEN_EXISTING,
-    FILE_ATTRIBUTE_NORMAL | FILE_FLAG_BACKUP_SEMANTICS,
-    NULL
-  );
-  if (file == INVALID_HANDLE_VALUE) {
-    *err_out = GetLastError();
-    return -1;
-  }
-
-  DWORD len = GetFinalPathNameByHandleW(
-    file,
-    job->buf,
-    REALPATH_JOB_BUFFER_LENGTH,
-    FILE_NAME_NORMALIZED | VOLUME_NAME_DOS
-  );
-
-  if (!len) {
-    *err_out = GetLastError();
-    CloseHandle(file);
-    return -1;
-  } else if (len < REALPATH_JOB_BUFFER_LENGTH) {
-    job->result = (char*)job->buf;
-  } else {
-    // include the extra NUL terminator
-    job->result = malloc((len + 1) * sizeof(WCHAR));
-    len = GetFinalPathNameByHandleW(
-      file,
-      (LPWSTR)job->result,
-      len,
-      FILE_NAME_NORMALIZED | VOLUME_NAME_DOS
-    );
-    if (!len) {
-      *err_out = GetLastError();
-      CloseHandle(file);
-      return -1;
-    }
-  }
-
-  CloseHandle(file);
-  return 0;
-
-#else
-
-  job->result = realpath(job->path, 0);
-  if (!job->result) {
-    *err_out = errno;
-    return -1;
-  }
-  return 0;
-
-#endif
-}
-
-struct realpath_job *moonbitlang_async_make_realpath_job(char *path) {
-  struct realpath_job *job = MAKE_JOB(realpath, 0);
-  job->path = path;
-  job->result = 0;
-  return job;
-}
-
-char *moonbitlang_async_get_realpath_result(struct realpath_job *job) {
-#ifdef _WIN32
-  if (wcsncmp((WCHAR*)job->result, L"\\\\?\\UNC\\", 8) == 0) {
-    ((WCHAR*)job->result)[6] = L'\\';
-    return job->result + 6 * sizeof(WCHAR);
-  }
-  else if (wcsncmp((WCHAR*)job->result, L"\\\\?\\", 4) == 0)
-    return job->result + 4 * sizeof(WCHAR);
-  else
-    return job->result + 8;
-#else
-  return job->result;
-#endif
 }
 
 // ===== spawn job, spawn foreign process =====
@@ -2771,55 +1498,3 @@ struct sigwait_job *moonbitlang_async_make_sigwait_job(int *signals) {
 }
 
 #endif // #ifndef _WIN32, sigwait job
-
-#ifndef _WIN32
-
-// ===== inotify_add_watch job, add path to watch with inotify =====
-struct inotify_add_watch_job {
-  HANDLE inotify;
-  char *path;
-  int32_t is_dir;
-};
-
-static
-void free_inotify_add_watch_job(struct inotify_add_watch_job *job) {
-  moonbit_decref(job->path);
-}
-
-static
-int32_t inotify_add_watch_job_worker(struct inotify_add_watch_job *job, int32_t *err_out) {
-#ifdef __linux__
-
-  uint32_t flags = job->is_dir
-    ? IN_CREATE | IN_MOVED_FROM | IN_MOVED_TO | IN_DELETE
-    : IN_MODIFY;
-
-  int32_t ret = inotify_add_watch(job->inotify, job->path, flags);
-  if (ret < 0)
-    *err_out = errno;
-  return ret;
-
-#else
-
-  *err_out = ENOSYS;
-  return 0;
-
-#endif
-}
-
-MOONBIT_FFI_EXPORT
-struct inotify_add_watch_job *moonbitlang_async_make_inotify_add_watch_job(
-  HANDLE inotify,
-  char *path,
-  int32_t is_dir
-) {
-  struct inotify_add_watch_job *job = MAKE_JOB(inotify_add_watch, 0);
-
-  job->inotify = inotify;
-  job->path = path;
-  job->is_dir = is_dir;
-
-  return job;
-}
-
-#endif // #ifndef _WIN32, `inotify_add_watch` job
